+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/crypto/verification/DefaultOutgoingSASVerificationRequest.kt
@@ -21,34 +21,22 @@ import im.vector.matrix.android.api.session.crypto.sas.OutgoingSasVerificationRe
 import im.vector.matrix.android.api.session.crypto.sas.SasVerificationTxState
 import im.vector.matrix.android.api.session.events.model.EventType
 import im.vector.matrix.android.internal.crypto.actions.SetDeviceVerificationAction
-import im.vector.matrix.android.internal.crypto.model.rest.KeyVerificationAccept
-import im.vector.matrix.android.internal.crypto.model.rest.KeyVerificationKey
-import im.vector.matrix.android.internal.crypto.model.rest.KeyVerificationMac
 import im.vector.matrix.android.internal.crypto.model.rest.KeyVerificationStart
 import im.vector.matrix.android.internal.crypto.store.IMXCryptoStore
-import im.vector.matrix.android.internal.crypto.tasks.SendToDeviceTask
-import im.vector.matrix.android.internal.task.TaskExecutor
-import im.vector.matrix.android.internal.util.JsonCanonicalizer
 import timber.log.Timber
 
-internal class OutgoingSASVerificationRequest(
-        private val sasVerificationService: DefaultSasVerificationService,
-        private val setDeviceVerificationAction: SetDeviceVerificationAction,
-        private val credentials: Credentials,
-        private val cryptoStore: IMXCryptoStore,
-        private val sendToDeviceTask: SendToDeviceTask,
-        private val taskExecutor: TaskExecutor,
+internal class DefaultOutgoingSASVerificationRequest(
+        setDeviceVerificationAction: SetDeviceVerificationAction,
+        credentials: Credentials,
+        cryptoStore: IMXCryptoStore,
         deviceFingerprint: String,
         transactionId: String,
         otherUserId: String,
-        otherDeviceId: String)
-    : SASVerificationTransaction(
-        sasVerificationService,
+        otherDeviceId: String
+) : SASVerificationTransaction(
         setDeviceVerificationAction,
         credentials,
         cryptoStore,
-        sendToDeviceTask,
-        taskExecutor,
         deviceFingerprint,
         transactionId,
         otherUserId,
@@ -78,26 +66,27 @@ internal class OutgoingSASVerificationRequest(
             }
         }
 
-    override fun onVerificationStart(startReq: KeyVerificationStart) {
-        Timber.e("## onVerificationStart - unexpected id:$transactionId")
+    override fun onVerificationStart(startReq: VerificationInfoStart) {
+        Timber.e("## SAS O: onVerificationStart - unexpected id:$transactionId")
         cancel(CancelCode.UnexpectedMessage)
     }
 
     fun start() {
         if (state != SasVerificationTxState.None) {
-            Timber.e("## start verification from invalid state")
+            Timber.e("## SAS O: start verification from invalid state")
             // should I cancel??
             throw IllegalStateException("Interactive Key verification already started")
         }
 
-        val startMessage = KeyVerificationStart()
-        startMessage.fromDevice = credentials.deviceId
-        startMessage.method = KeyVerificationStart.VERIF_METHOD_SAS
-        startMessage.transactionID = transactionId
-        startMessage.keyAgreementProtocols = KNOWN_AGREEMENT_PROTOCOLS
-        startMessage.hashes = KNOWN_HASHES
-        startMessage.messageAuthenticationCodes = KNOWN_MACS
-        startMessage.shortAuthenticationStrings = KNOWN_SHORT_CODES
+        val startMessage = transport.createStart(
+                credentials.deviceId ?: "",
+                KeyVerificationStart.VERIF_METHOD_SAS,
+                transactionId,
+                KNOWN_AGREEMENT_PROTOCOLS,
+                KNOWN_HASHES,
+                KNOWN_MACS,
+                KNOWN_SHORT_CODES
+        )
 
         startReq = startMessage
         state = SasVerificationTxState.SendingStart
@@ -111,10 +100,33 @@ internal class OutgoingSASVerificationRequest(
         )
     }
 
-    override fun onVerificationAccept(accept: KeyVerificationAccept) {
-        Timber.v("## onVerificationAccept id:$transactionId")
+//    fun request() {
+//        if (state != SasVerificationTxState.None) {
+//            Timber.e("## start verification from invalid state")
+//            // should I cancel??
+//            throw IllegalStateException("Interactive Key verification already started")
+//        }
+//
+//        val requestMessage = KeyVerificationRequest(
+//                fromDevice = session.sessionParams.credentials.deviceId ?: "",
+//                methods = listOf(KeyVerificationStart.VERIF_METHOD_SAS),
+//                timestamp = System.currentTimeMillis().toInt(),
+//                transactionID = transactionId
+//        )
+//
+//        sendToOther(
+//                EventType.KEY_VERIFICATION_REQUEST,
+//                requestMessage,
+//                SasVerificationTxState.None,
+//                CancelCode.User,
+//                null
+//        )
+//    }
+
+    override fun onVerificationAccept(accept: VerificationInfoAccept) {
+        Timber.v("## SAS O: onVerificationAccept id:$transactionId")
         if (state != SasVerificationTxState.Started) {
-            Timber.e("## received accept request from invalid state $state")
+            Timber.e("## SAS O: received accept request from invalid state $state")
             cancel(CancelCode.UnexpectedMessage)
             return
         }
@@ -123,7 +135,7 @@ internal class OutgoingSASVerificationRequest(
                 || !KNOWN_HASHES.contains(accept.hash)
                 || !KNOWN_MACS.contains(accept.messageAuthenticationCode)
                 || accept.shortAuthenticationStrings!!.intersect(KNOWN_SHORT_CODES).isEmpty()) {
-            Timber.e("## received accept request from invalid state")
+            Timber.e("## SAS O: received accept request from invalid state")
             cancel(CancelCode.UnknownMethod)
             return
         }
@@ -137,7 +149,7 @@ internal class OutgoingSASVerificationRequest(
         // and replies with a to_device message with type set to “m.key.verification.key”, sending Alice’s public key QA
         val pubKey = getSAS().publicKey
 
-        val keyToDevice = KeyVerificationKey.create(transactionId, pubKey)
+        val keyToDevice = transport.createKey(transactionId, pubKey)
         // we need to send this to other device now
         state = SasVerificationTxState.SendingKey
         sendToOther(EventType.KEY_VERIFICATION_KEY, keyToDevice, SasVerificationTxState.KeySent, CancelCode.User) {
@@ -148,8 +160,8 @@ internal class OutgoingSASVerificationRequest(
         }
     }
 
-    override fun onKeyVerificationKey(userId: String, vKey: KeyVerificationKey) {
-        Timber.v("## onKeyVerificationKey id:$transactionId")
+    override fun onKeyVerificationKey(userId: String, vKey: VerificationInfoKey) {
+        Timber.v("## SAS O: onKeyVerificationKey id:$transactionId")
         if (state != SasVerificationTxState.SendingKey && state != SasVerificationTxState.KeySent) {
             Timber.e("## received key from invalid state $state")
             cancel(CancelCode.UnexpectedMessage)
@@ -163,7 +175,7 @@ internal class OutgoingSASVerificationRequest(
         // in Bob’s m.key.verification.key and the content of Alice’s m.key.verification.start message.
 
         // check commitment
-        val concat = vKey.key + JsonCanonicalizer.getCanonicalJson(KeyVerificationStart::class.java, startReq!!)
+        val concat = vKey.key + startReq!!.toCanonicalJson()
         val otherCommitment = hashUsingAgreedHashMethod(concat) ?: ""
 
         if (accepted!!.commitment.equals(otherCommitment)) {
@@ -190,14 +202,14 @@ internal class OutgoingSASVerificationRequest(
         }
     }
 
-    override fun onKeyVerificationMac(vKey: KeyVerificationMac) {
-        Timber.v("## onKeyVerificationMac id:$transactionId")
+    override fun onKeyVerificationMac(vKey: VerificationInfoMac) {
+        Timber.v("## SAS O: onKeyVerificationMac id:$transactionId")
         if (state != SasVerificationTxState.OnKeyReceived
                 && state != SasVerificationTxState.ShortCodeReady
                 && state != SasVerificationTxState.ShortCodeAccepted
                 && state != SasVerificationTxState.SendingMac
                 && state != SasVerificationTxState.MacSent) {
-            Timber.e("## received key from invalid state $state")
+            Timber.e("## SAS O: received key from invalid state $state")
             cancel(CancelCode.UnexpectedMessage)
             return
         }