+++ vector/src/main/java/im/vector/riotx/features/command/Command.kt
@@ -38,7 +38,10 @@ enum class Command(val command: String, val parameters: String, @StringRes val d
     CHANGE_DISPLAY_NAME("/nick", "<display-name>", R.string.command_description_nick),
     MARKDOWN("/markdown", "<on|off>", R.string.command_description_markdown),
     CLEAR_SCALAR_TOKEN("/clear_scalar_token", "", R.string.command_description_clear_scalar_token),
-    SPOILER("/spoiler", "<message>", R.string.command_description_spoiler);
+    SPOILER("/spoiler", "<message>", R.string.command_description_spoiler),
+    SHRUG("/shrug", "<message>", R.string.command_description_shrug),
+    // TODO temporary command
+    VERIFY_USER("/verify", "<user-id>", R.string.command_description_verify);
 
     val length
         get() = command.length + 1